requirements.txt
fastapi
uvicorn
psycopg2-binary
pyTelegramBotAPI
requests
pytest
httpx
